+++ crypto/signatures_test.go
@@ -1,4 +1,4 @@
-package signatures
+package crypto
 
 import (
 	"crypto/rand"
